code integration fixes
git-svn-id: https://svn.ee.ucl.ac.uk/repos/jefferson/userspacerouter@1340 83772449-cf8d-4b79-a1b1-a7e526131002

+++ extras/IKMS/src/ikms/operations/InformationFlowConfigurationAndStatisticsOperation.java
@@ -459,9 +459,9 @@ public class InformationFlowConfigurationAndStatisticsOperation {
 		//}
 			
 			Float a;
-			a=0.0;
+			a=(float) 0.0;
 			a=0.0f;
-			if (a>0.0) a=1;
+			if (a>0.0) a=(float) 1;
 			
 			
 		return 0.0;
